requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
pandas==2.2.2
Pillow==10.4.0
matplotlib==3.9.1
scikit-learn==1.5.1
tqdm==4.66.4
reportlab==4.2.2
